[KAIZEN-0] bytte ut useFetch for react-query for innloggetsaksbehandler
react-query er mer brukt, aktivt forvaltet og støtter refetching mye mer elegant.
