+++ src/app/FetchSessionInfoOgLeggIRedux.tsx
@@ -1,7 +1,7 @@
 import { useInitializeLogger } from '../utils/logger/frontendLogger';
 import gsaktemaResource from '../rest/resources/gsaktemaResource';
 import baseurls from '../rest/resources/baseurlsResource';
-import innloggetSaksbehandler from '../rest/resources/innloggetSaksbehandler';
+import innloggetSaksbehandler from '../rest/resources/innloggetSaksbehandlerResource';
 import saksbehandlersEnheter from '../rest/resources/saksbehandlersEnheter';
 import { useQueryClient } from '@tanstack/react-query';
